update: Extract types, constants, and utils from KanbanBoard
